Mode demo ito app module
