+++ app/build.gradle
@@ -1,6 +1,7 @@
 plugins {
     id 'com.android.application'
     id 'kotlin-android'
+    id 'kotlin-android-extensions'
 }
 
 android {
@@ -9,7 +10,7 @@ android {
 
     defaultConfig {
         applicationId "com.verygoodsecurity.androiddemo"
-        minSdkVersion 19
+        minSdkVersion 21
         targetSdkVersion 30
         versionCode 1
         versionName "1.0"
@@ -33,6 +34,13 @@ android {
 }
 
 dependencies {
+
+    //VGS Collect SDK
+    implementation 'com.verygoodsecurity:vgscollect:1.6.1'
+    implementation 'com.verygoodsecurity:adapter-bouncer:1.0.3'
+
+    //VGS Show SDK
+    implementation 'com.verygoodsecurity:vgsshow:1.1.0'
 
     implementation "org.jetbrains.kotlin:kotlin-stdlib:$kotlin_version"
     implementation 'androidx.core:core-ktx:1.3.2'
