+++ app/src/main/res/drawable/bg_card_preview.xml
@@ -1,0 +1,9 @@
+<?xml version="1.0" encoding="utf-8"?>
+<shape xmlns:android="http://schemas.android.com/apk/res/android">
+    <corners android:radius="@dimen/half_default_padding"/>
+    <gradient
+        android:type="linear"
+        android:angle="-45"
+        android:startColor="@color/bg_card_preview_start_color"
+        android:endColor="@color/bg_card_preview_start_end"/>
+</shape>
