+++ app/src/main/res/drawable/ic_attach_file.xml
@@ -1,0 +1,10 @@
+<?xml version="1.0" encoding="utf-8"?>
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="8dp"
+    android:height="16dp"
+    android:viewportWidth="8"
+    android:viewportHeight="16">
+    <path
+        android:pathData="M6.9091,3.6364V12C6.9091,12.7715 6.6026,13.5115 6.057,14.057C5.5115,14.6026 4.7715,14.9091 4,14.9091C3.2285,14.9091 2.4885,14.6026 1.943,14.057C1.3974,13.5115 1.0909,12.7715 1.0909,12V2.9091C1.0909,2.4269 1.2825,1.9644 1.6234,1.6234C1.9644,1.2825 2.4269,1.0909 2.9091,1.0909C3.3913,1.0909 3.8538,1.2825 4.1947,1.6234C4.5357,1.9644 4.7273,2.4269 4.7273,2.9091V10.5455C4.7273,10.7383 4.6507,10.9233 4.5143,11.0597C4.3779,11.1961 4.1929,11.2727 4,11.2727C3.8071,11.2727 3.6221,11.1961 3.4857,11.0597C3.3493,10.9233 3.2727,10.7383 3.2727,10.5455V3.6364H2.1818V10.5455C2.1818,11.0277 2.3734,11.4901 2.7143,11.8311C3.0553,12.1721 3.5178,12.3636 4,12.3636C4.4822,12.3636 4.9447,12.1721 5.2856,11.8311C5.6266,11.4901 5.8182,11.0277 5.8182,10.5455V2.9091C5.8182,2.1376 5.5117,1.3976 4.9661,0.8521C4.4206,0.3065 3.6806,0 2.9091,0C2.1376,0 1.3976,0.3065 0.8521,0.8521C0.3065,1.3976 0,2.1376 0,2.9091V12C0,13.0609 0.4214,14.0783 1.1716,14.8284C1.9217,15.5786 2.9391,16 4,16C5.0609,16 6.0783,15.5786 6.8284,14.8284C7.5786,14.0783 8,13.0609 8,12V3.6364H6.9091Z"
+        android:fillColor="#1F8AB0"/>
+</vector>
