+++ app/src/main/res/drawable/ic_cards.xml
@@ -1,0 +1,10 @@
+<?xml version="1.0" encoding="utf-8"?>
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="24dp"
+    android:height="18dp"
+    android:viewportWidth="24"
+    android:viewportHeight="18">
+    <path
+        android:pathData="M21.8182,6V3H6.5455V6H21.8182ZM21.8182,0C22.3968,0 22.9518,0.2107 23.361,0.5858C23.7701,0.9609 24,1.4696 24,2V12C24,12.5304 23.7701,13.0391 23.361,13.4142C22.9518,13.7893 22.3968,14 21.8182,14H6.5455C5.9668,14 5.4118,13.7893 5.0027,13.4142C4.5935,13.0391 4.3636,12.5304 4.3636,12V2C4.3636,1.4696 4.5935,0.9609 5.0027,0.5858C5.4118,0.2107 5.9668,0 6.5455,0H21.8182ZM2.1818,16H18.5455V18H2.1818C1.6032,18 1.0482,17.7893 0.639,17.4142C0.2299,17.0391 0,16.5304 0,16V5H2.1818V16Z"
+        android:fillColor="#3B495C"/>
+</vector>
