+++ app/src/main/res/drawable/ic_custom_visa.xml
@@ -1,0 +1,18 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="50dp"
+    android:height="16dp"
+    android:viewportWidth="50"
+    android:viewportHeight="16">
+  <path
+      android:pathData="M3.0913,3.2232L4.9569,15.7827H9.5376L18.0784,0.231H13.265L9.6384,7.4437C9.1265,8.4678 8.6804,9.3936 8.2964,10.2428C8.2266,9.3572 8.149,8.497 8.056,7.6186L7.3229,0.2346H0L7.6138,5.8947L3.0913,3.2232Z"
+      android:fillColor="#EBEFF5"/>
+  <path
+      android:pathData="M21.3682,15.7851L24.506,0.2334H20.1969L17.0552,15.7851H21.3682Z"
+      android:fillColor="#EBEFF5"/>
+  <path
+      android:pathData="M23.8922,14.8665C24.9006,15.5772 26.677,16 28.6435,16C32.72,16 35.4583,13.8351 35.4583,10.6169C35.4583,8.8565 34.3645,7.4606 32.1188,6.3563C30.8621,5.6929 30.098,5.2483 30.098,4.687C30.098,4.1622 30.6837,3.5973 32.0722,3.5973C33.4259,3.5973 34.4576,4.0966 34.6554,4.195L35.1441,4.4465L36.5211,0.9841L36.1603,0.7945C35.1286,0.2588 33.8487,0 32.2468,0C28.3449,0 25.6182,2.143 25.6182,5.2118C25.6182,7.4351 27.4799,8.6706 29.1167,9.5016C30.6527,10.2524 30.9552,10.6788 30.9552,11.1636C30.9552,12.0966 29.8847,12.4282 28.9654,12.4282C27.7669,12.4282 26.518,12.1112 25.5367,11.5572L25.0519,11.2875L23.5935,14.6624L23.8922,14.8665Z"
+      android:fillColor="#EBEFF5"/>
+  <path
+      android:pathData="M39.6168,15.7851L41.4669,11.9837H45.1982L45.524,15.7851H50L48.0607,0.2334H43.0533L35.0012,15.7851H39.6168ZM44.1161,6.1778C44.2867,5.8279 44.4613,5.427 44.6397,5.0152C44.6707,5.3833 44.6979,5.7295 44.7211,6.0284L44.9733,8.6853H42.8788L44.1161,6.1778Z"
+      android:fillColor="#EBEFF5"/>
+</vector>
