+++ app/src/main/res/drawable/ic_scan_card.xml
@@ -1,0 +1,10 @@
+<?xml version="1.0" encoding="utf-8"?>
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="24dp"
+    android:height="20dp"
+    android:viewportWidth="24"
+    android:viewportHeight="20">
+    <path
+        android:pathData="M18,4H6C5.4696,4 4.9609,4.2107 4.5858,4.5858C4.2107,4.9609 4,5.4696 4,6V14C4,14.5304 4.2107,15.0391 4.5858,15.4142C4.9609,15.7893 5.4696,16 6,16H18C18.5304,16 19.0391,15.7893 19.4142,15.4142C19.7893,15.0391 20,14.5304 20,14V6C20,5.4696 19.7893,4.9609 19.4142,4.5858C19.0391,4.2107 18.5304,4 18,4ZM18,10H6V7H18V10ZM2,2H6V0H2C1.4696,0 0.9609,0.2107 0.5858,0.5858C0.2107,0.9609 0,1.4696 0,2L0,6H2V2ZM22,0H18V2H22V6H24V2C24,1.4696 23.7893,0.9609 23.4142,0.5858C23.0391,0.2107 22.5304,0 22,0ZM2,14H0V18C0,18.5304 0.2107,19.0391 0.5858,19.4142C0.9609,19.7893 1.4696,20 2,20H6V18H2V14ZM22,18H18V20H22C22.5304,20 23.0391,19.7893 23.4142,19.4142C23.7893,19.0391 24,18.5304 24,18V14H22V18Z"
+        android:fillColor="#ffffff"/>
+</vector>
