+++ app/src/main/res/layout/activity_payment_card_number_field.xml
@@ -1,0 +1,24 @@
+<?xml version="1.0" encoding="utf-8"?>
+<LinearLayout
+    xmlns:android="http://schemas.android.com/apk/res/android"
+    android:id="@+id/parentLayout"
+    android:layout_width="match_parent"
+    android:layout_height="match_parent"
+    android:orientation="vertical"
+    android:padding="@dimen/default_padding"
+    xmlns:app="http://schemas.android.com/apk/res-auto">
+
+    <com.verygoodsecurity.vgscollect.widget.VGSCardNumberEditText
+        android:id="@+id/cardNumberFromLayout"
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        android:layout_marginTop="@dimen/default_padding"
+        android:layout_marginBottom="@dimen/default_padding"
+        app:fieldName="card_number_field_name"
+        app:numberDivider="-"
+        app:enableValidation="true"
+        app:cardBrandIconGravity="end"
+        app:validationRule="acceptUnknown"
+        app:hint="cardNumber"/>
+
+</LinearLayout>
