+++ app/src/main/res/layout/activity_states.xml
@@ -1,0 +1,15 @@
+<?xml version="1.0" encoding="utf-8"?>
+<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    android:layout_width="match_parent"
+    android:layout_height="match_parent"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    android:orientation="vertical">
+
+    <com.verygoodsecurity.vgscollect.widget.VGSCardNumberEditText
+        android:id="@+id/field"
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        app:cardBrandIconGravity="end"
+        app:numberDivider="-"/>
+
+</LinearLayout>
