+++ app/src/main/res/layout/layout_collect_payment_card.xml
