+++ app/src/main/res/layout/layout_preview_card.xml
