+++ app/src/main/res/layout/layout_show_reveal_card.xml
