+++ app/src/main/res/menu/scan_menu.xml
@@ -1,0 +1,8 @@
+<?xml version="1.0" encoding="utf-8"?>
+<menu xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto">
+    <item android:id="@+id/scan_card"
+        android:icon="@drawable/ic_scan_card"
+        android:title="@string/scan_item"
+        app:showAsAction="always"/>
+</menu>
