+++ app/src/main/res/values/colors.xml
@@ -1,10 +1,19 @@
 <?xml version="1.0" encoding="utf-8"?>
 <resources>
-    <color name="purple_200">#FFBB86FC</color>
-    <color name="purple_500">#FF6200EE</color>
-    <color name="purple_700">#FF3700B3</color>
-    <color name="teal_200">#FF03DAC5</color>
-    <color name="teal_700">#FF018786</color>
-    <color name="black">#FF000000</color>
-    <color name="white">#FFFFFFFF</color>
+    <color name="colorPrimary">#27A386</color>
+    <color name="colorPrimaryDark">#188269</color>
+    <color name="colorAccent">#26A6D4</color>
+
+    <color name="title_color">#3B495C</color>
+
+    <color name="bg_card_preview_start_color">#0C1427</color>
+    <color name="bg_card_preview_start_end">#2f426f</color>
+
+    <color name="preview_holder_name_color">#CED5E0</color>
+
+    <color name="bg_vgs_card_view">#F7F9FC</color>
+
+    <color name="state_unactive">#C5E1EB</color>
+    <color name="state_active">@color/colorAccent</color>
+
 </resources>
