+++ app/src/main/res/values/dimens.xml
@@ -1,0 +1,28 @@
+<?xml version="1.0" encoding="utf-8"?>
+<resources>
+    <dimen name="default_padding">16dp</dimen>
+    <dimen name="half_default_padding">8dp</dimen>
+    <dimen name="button_size">48dp</dimen>
+
+
+    <dimen name="collect_activity_cards_title_text_size">20dp</dimen>
+    <dimen name="field_wrapper_height">52dp</dimen>
+
+    <dimen name="preview_holder_name_text_size">16dp</dimen>
+    <dimen name="preview_card_horizontal_padding">40dp</dimen>
+    <dimen name="preview_card_vertical_padding">27dp</dimen>
+
+    <dimen name="payment_card_horizontal_margin">20dp</dimen>
+
+    <dimen name="payment_card_horizontal_padding">20dp</dimen>
+    <dimen name="payment_card_vertical_padding">20dp</dimen>
+
+    <dimen name="vgs_field_horizontal_padding">21dp</dimen>
+    <dimen name="vgs_field_vertical_padding">2dp</dimen>
+    <dimen name="vgs_field_text_size">14dp</dimen>
+
+    <dimen name="password_icon_size">48dp</dimen>
+
+    <dimen name="card_preview_width">300dp</dimen>
+    <dimen name="card_preview_height">190dp</dimen>
+</resources>
